Update and add tests.
