+++ tests/ocean/test_ocean_agreements.py
@@ -14,6 +14,9 @@ from squid_py.keeper.web3_provider import Web3Provider
 from squid_py.ocean.ocean_agreements import OceanAgreements
 from tests.resources.helper_functions import (get_ddo_sample, log_event)
 from tests.resources.tiers import e2e_test
+from squid_py.agreements.service_agreement import ServiceAgreement
+from squid_py.brizo.brizo import Brizo
+from tests.resources.mocks.brizo_mock import BrizoMock
 
 
 @pytest.fixture
@@ -159,3 +162,116 @@ def test_agreement_status(setup_agreements_enviroment, ocean_agreements):
                                                                     "escrowReward": 2
                                                                     }
                                                      }
+
+
+@e2e_test
+def test_sign_agreement(publisher_ocean_instance, consumer_ocean_instance, registered_ddo):
+    # point consumer_ocean_instance's brizo mock to the publisher's ocean instance
+    Brizo.set_http_client(
+        BrizoMock(publisher_ocean_instance, publisher_ocean_instance.main_account))
+
+    consumer_ocn = consumer_ocean_instance
+    consumer_acc = consumer_ocn.main_account
+    keeper = Keeper.get_instance()
+
+    pub_ocn = publisher_ocean_instance
+    publisher_acc = pub_ocn.main_account
+
+    service_definition_id = '1'
+    did = registered_ddo.did
+    asset_id = registered_ddo.asset_id
+    ddo = consumer_ocn.assets.resolve(did)
+    service_agreement = ServiceAgreement.from_ddo(service_definition_id, ddo)
+    price = service_agreement.get_price()
+
+    # Give consumer some tokens
+    keeper.dispenser.request_tokens(price * 2, consumer_acc)
+
+    agreement_id, signature = consumer_ocean_instance.agreements.prepare(
+        did, service_agreement.service_definition_id, consumer_acc)
+
+    success = publisher_ocean_instance.agreements.create(
+        did,
+        service_agreement.service_definition_id,
+        agreement_id,
+        signature,
+        consumer_acc.address,
+        publisher_acc
+    )
+    assert success, 'createAgreement failed.'
+
+    event = keeper.escrow_access_secretstore_template.subscribe_agreement_created(
+        agreement_id,
+        10,
+        log_event(keeper.escrow_access_secretstore_template.AGREEMENT_CREATED_EVENT),
+        (),
+        wait=True
+    )
+    assert event, 'no event for AgreementCreated '
+
+    # Verify condition types (condition contracts)
+    agreement_values = keeper.agreement_manager.get_agreement(agreement_id)
+    assert agreement_values.did == asset_id, ''
+    cond_types = keeper.escrow_access_secretstore_template.get_condition_types()
+    for i, cond_id in enumerate(agreement_values.condition_ids):
+        cond = keeper.condition_manager.get_condition(cond_id)
+        assert cond.type_ref == cond_types[i]
+        assert int(cond.state) == 1
+
+    access_cond_id, lock_cond_id, escrow_cond_id = agreement_values.condition_ids
+    # Fulfill lock_reward_condition
+    starting_balance = keeper.token.get_token_balance(keeper.escrow_reward_condition.address)
+    keeper.token.token_approve(keeper.lock_reward_condition.address, price, consumer_acc)
+    tx_hash = keeper.lock_reward_condition.fulfill(
+        agreement_id, keeper.escrow_reward_condition.address, price, consumer_acc
+    )
+    keeper.lock_reward_condition.get_tx_receipt(tx_hash)
+    assert keeper.token.get_token_balance(
+        keeper.escrow_reward_condition.address) == (price + starting_balance), ''
+    assert keeper.condition_manager.get_condition_state(lock_cond_id) == 2, ''
+    event = keeper.lock_reward_condition.subscribe_condition_fulfilled(
+        agreement_id,
+        10,
+        log_event(keeper.lock_reward_condition.FULFILLED_EVENT),
+        (),
+        wait=True
+    )
+    assert event, 'no event for LockRewardCondition.Fulfilled'
+
+    # Fulfill access_secret_store_condition
+    tx_hash = keeper.access_secret_store_condition.fulfill(
+        agreement_id, asset_id, consumer_acc.address, publisher_acc
+    )
+    keeper.access_secret_store_condition.get_tx_receipt(tx_hash)
+    assert 2 == keeper.condition_manager.get_condition_state(access_cond_id), ''
+    event = keeper.access_secret_store_condition.subscribe_condition_fulfilled(
+        agreement_id,
+        10,
+        log_event(keeper.access_secret_store_condition.FULFILLED_EVENT),
+        (),
+        wait=True
+    )
+    assert event, 'no event for AccessSecretStoreCondition.Fulfilled'
+
+    # Fulfill escrow_reward_condition
+    tx_hash = keeper.escrow_reward_condition.fulfill(
+        agreement_id, price, publisher_acc.address,
+        consumer_acc.address, lock_cond_id,
+        access_cond_id, publisher_acc
+    )
+    keeper.escrow_reward_condition.get_tx_receipt(tx_hash)
+    assert 2 == keeper.condition_manager.get_condition_state(escrow_cond_id), ''
+    event = keeper.escrow_reward_condition.subscribe_condition_fulfilled(
+        agreement_id,
+        10,
+        log_event(keeper.escrow_reward_condition.FULFILLED_EVENT),
+        (),
+        wait=True
+    )
+    assert event, 'no event for EscrowReward.Fulfilled'
+
+    # path = consumer_ocean_instance.assets.consume(
+    #     agreement_id, did, service_definition_id,
+    #     consumer_acc, ConfigProvider.get_config().downloads_path
+    # )
+    # print('All good, files are here: %s' % path)
