+++ tests/ocean/test_ocean_assets.py
@@ -1,3 +1,71 @@
 #  Copyright 2018 Ocean Protocol Foundation
 #  SPDX-License-Identifier: Apache-2.0
 
+import logging
+
+import pytest
+
+from squid_py.ddo.ddo import DDO
+from squid_py.ddo.metadata import Metadata
+from squid_py.did import DID
+from squid_py.exceptions import OceanDIDNotFound
+from tests.resources.helper_functions import get_resource_path
+from tests.resources.tiers import e2e_test
+
+
+@e2e_test
+def test_register_asset(publisher_ocean_instance):
+    logging.debug("".format())
+    sample_ddo_path = get_resource_path('ddo', 'ddo_sa_sample.json')
+    assert sample_ddo_path.exists(), "{} does not exist!".format(sample_ddo_path)
+
+    ##########################################################
+    # Setup account
+    ##########################################################
+    publisher = publisher_ocean_instance.main_account
+
+    # ensure Ocean token balance
+    if publisher_ocean_instance.accounts.balance(publisher).ocn == 0:
+        publisher_ocean_instance.accounts.request_tokens(publisher, 200)
+
+    # You will need some token to make this transfer!
+    assert publisher_ocean_instance.accounts.balance(publisher).ocn > 0
+
+    ##########################################################
+    # Create an asset DDO with valid metadata
+    ##########################################################
+    asset = DDO(json_filename=sample_ddo_path)
+
+    ##########################################################
+    # Register using high-level interface
+    ##########################################################
+    publisher_ocean_instance.assets.create(asset.metadata, publisher)
+
+
+@e2e_test
+def test_resolve_did(publisher_ocean_instance):
+    # prep ddo
+    metadata = Metadata.get_example()
+    publisher = publisher_ocean_instance.main_account
+    original_ddo = publisher_ocean_instance.assets.create(metadata, publisher)
+
+    # happy path
+    did = original_ddo.did
+    ddo = publisher_ocean_instance.assets.resolve(did).as_dictionary()
+    original = original_ddo.as_dictionary()
+    assert ddo['publicKey'] == original['publicKey']
+    assert ddo['authentication'] == original['authentication']
+    assert ddo['service']
+    assert original['service']
+    # assert ddo['service'][:-1] == original['service'][:-1]
+    # assert ddo == original_ddo.as_dictionary(), 'Resolved ddo does not match original.'
+
+    # Can't resolve unregistered asset
+    unregistered_did = DID.did()
+    with pytest.raises(OceanDIDNotFound):
+        publisher_ocean_instance.assets.resolve(unregistered_did)
+
+    # Raise error on bad did
+    invalid_did = "did:op:0123456789"
+    with pytest.raises(OceanDIDNotFound):
+        publisher_ocean_instance.assets.resolve(invalid_did)
